refactor: add heroSection folder in components
