+++ client/src/components/heroSection/heroStyles.tsx
