+++ client/src/pages/Landing/index.tsx
@@ -1,4 +1,4 @@
-import HeroSection from '../../components/heroSection';
+import HeroSection from '../../components/heroSection/heroSection';
 
 function Landing() {
   return (<HeroSection />);
